Add source files
Each packages in the source file, controller and view packages, have have a distinct function. The ContentCeckerOGK in the controller package checks the uploaded text for content while the EmailContentCheckerOGK, also in the controller package, implements the contentCheckerOGK, the FileContentCheckerOGK houses the function that reads the files, reads the text in the file from bytes to characters, from characters to texts. Then the FileContentCheckerGUIOGK, in the view package, house the GUI JFrame, JPanel, JButton, JTextFeild, and the UI needed in the GUI application.